chore: modify sync script to update based on git-ref
If provided the update-to-head.sh script will now accept an upstream
git-ref (typically a release tag, e.g. 1.2.2). If a git-ref is provided,
there MUST be a corresponding branch in securesign/rekor with the same
name, prefixed with "midstream-", e.g. "midstream-v1.2.2".

Finally, the script will also push to origin/redhat-<git-ref> if a
git-ref is provided.

This allows us to maintain branch history in midstream, corresponding to
upstream releases, and typically pins the upstream dockerfile shas.

Signed-off-by: Lance Ball <lball@redhat.com>

fixup

Signed-off-by: Lance Ball <lball@redhat.com>